requirements.txt
python-telegram-bot>=21.0
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
aiofiles==23.2.1
python-dotenv==1.0.0

